+++ packages/gui/src/components/html/editor.tsx
@@ -87,7 +87,9 @@ export function HtmlEditor({ onChange }: HtmlEditorProps) {
           }
           onRemove={() => {
             onChange(removeChildAtPath(value, selected))
-            setSelected(null)
+            const newPath = [...selected]
+            newPath.pop()
+            setSelected(newPath)
           }}
         />
       )}
@@ -119,7 +121,9 @@ function NodeEditor({ value, onChange, onRemove }: TagEditorProps) {
         minHeight: '100%',
       }}
     >
-      <div sx={{ mb: 2, display: 'flex', alignItems: 'flex-end', px: 3, pt: 3, }}>
+      <div
+        sx={{ mb: 2, display: 'flex', alignItems: 'flex-end', px: 3, pt: 3 }}
+      >
         <SelectInput
           label="Type"
           value={nodeType}
@@ -130,6 +134,7 @@ function NodeEditor({ value, onChange, onRemove }: TagEditorProps) {
               onChange({
                 type: 'element',
                 tagName: 'div',
+                children: [],
               })
             }
           }}
@@ -152,10 +157,10 @@ function NodeSwitch({ value, onChange }: EditorProps) {
       <div sx={{ px: 3 }}>
         <Label>
           <span sx={{ display: 'block' }}>Content</span>
-          <input
-            type='text'
-            sx={{ 
-              width: '100%'
+          <textarea
+            rows={8}
+            sx={{
+              width: '100%',
             }}
             value={value.value}
             onChange={(e) =>
@@ -180,7 +185,7 @@ function NodeSwitch({ value, onChange }: EditorProps) {
           pb: 3,
         }}
       >
-        <div sx={{ mb: 2, px: 3,  }}>
+        <div sx={{ mb: 2, px: 3 }}>
           <Label>Tag name</Label>{' '}
           <Combobox
             onFilterItems={(filterValue) => {
@@ -209,7 +214,7 @@ function NodeSwitch({ value, onChange }: EditorProps) {
           />
         </div>
       </article>
-      <div sx={{ pt: 3, px: 3  }}>
+      <div sx={{ pt: 3, px: 3 }}>
         <Label> 🎨 Styles</Label>
         <div sx={{ mt: 2 }}>
           <Editor
@@ -305,7 +310,7 @@ function TreeNode({ value, path, onSelect, onChange }: TreeNodeProps) {
                         value: '',
                       })
                     )
-                    onSelect(null)
+                    onSelect([i])
                   }}
                 />
                 <TreeNode