+++ packages/gui/src/components/inputs/KeywordInput.tsx
@@ -1,0 +1,25 @@
+import { EditorPropsWithLabel } from '../../types/editor'
+import { Label } from '../primitives'
+import { KeywordSelect } from '../primitives/KeywordSelect'
+
+interface Props<T extends string> extends EditorPropsWithLabel<T> {
+  options: T[]
+}
+export function KeywordInput<T extends string>(props: Props<T>) {
+  return (
+    <div>
+      <Label>{props.label}</Label>
+      <div
+        sx={{
+          border: '1px solid',
+          borderColor: 'border',
+          borderRadius: '0.25rem',
+          width: 'max-content',
+          px: 1,
+        }}
+      >
+        <KeywordSelect {...props} />
+      </div>
+    </div>
+  )
+}