+++ packages/gui/src/types/editor.ts
@@ -1,6 +1,11 @@
 export interface EditorProps<T> {
   value: T
   onChange(newValue: T): void
+  /**
+   * Indicates whether the input is a top-level input (ie. not a sub-input)
+   * and indicates it should be populated with things like global keywords.
+   */
+  topLevel?: boolean
   onRemove?(): void
 }